# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      parameters:
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── INSTRUMENTS ────────────────────────────────────────────────────────────
...
  /confluence-categories:
    get:
      operationId: listConfluenceCategories
      tags: [confluences]
      summary: List confluence categories with their confluences
      responses:
        "200":
          description: List of categories with confluences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConfluenceCategory"
    post:
      operationId: createConfluenceCategory
      tags: [confluences]
      summary: Create a confluence category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfluenceCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfluenceCategory"
  /confluence-categories/{id}:
    patch:
      operationId: updateConfluenceCategory
      tags: [confluences]
      summary: Update confluence category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfluenceCategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfluenceCategory"
        "404":
          description: Not found
    delete:
      operationId: deleteConfluenceCategory
      tags: [confluences]
      summary: Delete confluence category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── CONFLUENCES ─────────────────────────────────────────────────────────────
  /confluences:
    post:
      operationId: createConfluence
      tags: [confluences]
      summary: Create a confluence item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfluenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Confluence"
  /confluences/{id}:
    patch:
      operationId: updateConfluence
      tags: [confluences]
      summary: Update confluence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfluenceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Confluence"
        "404":
          description: Not found
    delete:
      operationId: deleteConfluence
      tags: [confluences]
      summary: Delete confluence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── TAGS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, name, baseCurrency, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        broker: { type: ["string", "null"] }
        accountType: { type: ["string", "null"] }
        baseCurrency: { type: string }
        startingBalance: { type: ["number", "null"] }
        currentBalance: { type: ["number", "null"] }
        defaultRiskPercentage: { type: ["number", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AccountInput:
      type: object
      required: [name, baseCurrency]
      properties:
        name: { type: string, minLength: 1 }
        broker: { type: string }
        accountType: { type: string }
        baseCurrency: { type: string, minLength: 1 }
        startingBalance: { type: number }
        currentBalance: { type: number }
        defaultRiskPercentage: { type: number }
        isActive: { type: boolean, default: true }
    AccountUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        broker: { type: string }
        accountType: { type: string }
        baseCurrency: { type: string }
        startingBalance: { type: ["number", "null"] }
        currentBalance: { type: ["number", "null"] }
        defaultRiskPercentage: { type: ["number", "null"] }
        isActive: { type: boolean }

    # ── INSTRUMENT ──────────────────────────────────────────────────────────
...
    ConfluenceCategory:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        confluences:
          type: array
          items:
            $ref: "#/components/schemas/Confluence"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ConfluenceCategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    ConfluenceCategoryUpdate:
      type: object
      properties:
        name: { type: string }

    # ── CONFLUENCE ───────────────────────────────────────────────────────────
    Confluence:
      type: object
      required: [id, categoryId, categoryName, name, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        categoryId: { type: integer }
        categoryName: { type: string }
        name: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ConfluenceInput:
      type: object
      required: [categoryId, name]
      properties:
        categoryId: { type: integer }
        name: { type: string, minLength: 1 }
        isActive: { type: boolean, default: true }
    ConfluenceUpdate:
      type: object
      properties:
        categoryId: { type: integer }
        name: { type: string }
        isActive: { type: boolean }

    # ── TAG ──────────────────────────────────────────────────────────────────
...