# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  broker?: string;
  accountType?: string;
  /** @minLength 1 */
  baseCurrency: string;
  startingBalance?: number;
  currentBalance?: number;
  defaultRiskPercentage?: number;
  isActive?: boolean;
}