# lib/api-zod/src/generated/types/confluence.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface Confluence {
  id: number;
  categoryId: number;
  categoryName: string;
  name: string;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}